Фикс скорости

+++ src/main/kotlin/me/kolotilov/letsagoservice/domain/models/Point.kt
@@ -37,7 +37,7 @@ fun List<Point>.duration(): Duration {
 }
 
 fun List<Point>.speed(): Double {
-    return (distance() / 1000) / (duration().millis / (60 * 60 * 1000))
+    return (distance() / 1000) / (duration().millis.toDouble() / (60 * 60 * 1000))
 }
 
 fun List<Point>.altitudeDelta(): Double {
